pyproject.toml
[tool.poetry]
name = "fund-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.2"
django-rest-framework = "^0.1.0"
black = "^24.2.0"
factory-boy = "^3.3.0"
arrow = "^1.3.0"

[tool.black]
line-length = 79
target-version = ['py310']
exclude = '''
/(
  | migrations
)/

'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
